move buffers to separate file, add delay buffer
