+++ robosuite/environments/manipulation/two_arm_handover.py
@@ -136,7 +136,7 @@ class TwoArmHandover(TwoArmEnv):
     def __init__(
         self,
         robots,
-        env_configuration="single-arm-opposed",
+        env_configuration="default",
         controller_configs=None,
         gripper_types="default",
         initialization_noise="default",
