+++ robosuite/environments/manipulation/two_arm_peg_in_hole.py
@@ -129,7 +129,7 @@ class TwoArmPegInHole(TwoArmEnv):
     def __init__(
         self,
         robots,
-        env_configuration="single-arm-opposed",
+        env_configuration="default",
         controller_configs=None,
         gripper_types=None,
         initialization_noise="default",
